Media Codec API just got a visit from the coding style police. We now use "const" for the chunk buffer pointer returned by GetNextChunk, because the buffer is not supposed to be modified by the codec. size_t is used for the size where applicable. This matches BMediaDecoder.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@13361 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/media/plugins/ac3_decoder/ac3_decoder.cpp
@@ -62,7 +62,7 @@ AC3Decoder::GetCodecInfo(media_codec_info *info)
 
 status_t
 AC3Decoder::Setup(media_format *ioEncodedFormat,
-				  const void *infoBuffer, int32 infoSize)
+				  const void *infoBuffer, size_t infoSize)
 {
 	return B_OK;
 }
